Split net library into client and server
