+++ lute/net/CMakeLists.txt
@@ -3,7 +3,8 @@ add_library(Lute.Net STATIC)
 target_sources(Lute.Net PRIVATE
     include/lute/net.h
 
-    src/net.cpp
+    src/client.cpp
+    src/server.cpp
 )
 
 target_compile_features(Lute.Net PUBLIC cxx_std_17)
